Renamed external library "headers" to "<headers>.
CMakePresets.json:
- added release configuration buidls.
- removed RTC_BUILD_BENCHMARKS cache variable from release configurations.

CMakeLists.txt:
- commented out address sanitizer block.

stream.h:
- fixed a warning treated as error (-Wconversion -Wsign-conversion).

+++ include/rtc/stream.h
@@ -5,6 +5,6 @@
 
 namespace rtc::stream {
     inline auto get_unread(std::istringstream& iss) {
-        return iss.eof() ? "" : iss.str().substr(iss.tellg());
+        return iss.eof() ? "" : iss.str().substr(static_cast<size_t>(iss.tellg()));
     }
 }  // namespace rtc::stream
